update web packages complying the new structure
